优化 CREATE TABLE 解析逻辑（解析逻辑改造）
