+++ metasequoia_sql/translate/__init__.py
@@ -1,2 +1,1 @@
-from metasequoia_sql.translate.from_mysql import *
 from metasequoia_sql.translate.to_hive import *
